+++ src/main/gen/org/jetbrains/lama/psi/api/LamaFunctionDefinition.java
@@ -4,11 +4,10 @@ package org.jetbrains.lama.psi.api;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiNameIdentifierOwner;
 import com.intellij.psi.StubBasedPsiElement;
 import org.jetbrains.lama.psi.stubs.LamaFunctionDefinitionStub;
 
-public interface LamaFunctionDefinition extends LamaControlFlowHolder, LamaDefinition, PsiNameIdentifierOwner, StubBasedPsiElement<LamaFunctionDefinitionStub> {
+public interface LamaFunctionDefinition extends LamaControlFlowHolder, LamaIdentifierOwnerDefinition, StubBasedPsiElement<LamaFunctionDefinitionStub> {
 
   @Nullable
   LamaFunctionBody getFunctionBody();