+++ src/main/gen/org/jetbrains/lama/psi/api/LamaInfixOperatorDefinition.java
@@ -4,12 +4,11 @@ package org.jetbrains.lama.psi.api;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiNamedElement;
 import com.intellij.psi.StubBasedPsiElement;
 import org.jetbrains.lama.psi.stubs.LamaInfixOperatorDefinitionStub;
 import org.jetbrains.lama.psi.stubs.LamaInfixAssociativity;
 
-public interface LamaInfixOperatorDefinition extends LamaControlFlowHolder, LamaDefinition, PsiNamedElement, StubBasedPsiElement<LamaInfixOperatorDefinitionStub> {
+public interface LamaInfixOperatorDefinition extends LamaControlFlowHolder, LamaDefinition, StubBasedPsiElement<LamaInfixOperatorDefinitionStub> {
 
   @Nullable
   LamaFunctionBody getFunctionBody();