+++ src/main/kotlin/org/jetbrains/lama/psi/api/LamaDefinition.kt
@@ -1,6 +1,11 @@
 package org.jetbrains.lama.psi.api
 
-interface LamaDefinition : LamaPsiElement {
+import com.intellij.psi.PsiNameIdentifierOwner
+import com.intellij.psi.PsiNamedElement
+
+interface LamaDefinition : LamaPsiElement, PsiNamedElement {
   val isPublic: Boolean
   val isTopLevel: Boolean
 }
+
+interface LamaIdentifierOwnerDefinition : LamaDefinition, PsiNameIdentifierOwner